+++ Kbuild
@@ -1,3 +1,3 @@
 obj-m += tmem_csc450.o
-tmem_csc450-y := src/main.o
-tmem_csc450-y += src/tmemscan.o
+tmem_csc450-y := src/tmem_main.o
+tmem_csc450-y += src/tmem_vmscan.o
